+++ src/main/scala/org/fs/checker/provider/impl/RuTracker.scala
@@ -37,7 +37,7 @@ class RuTracker(httpClient: HttpClient, override val dumpService: PageContentDum
   override def parseDateLastUpdated(content: String): TorrentParseResult = {
     // Format (bottom section): 10-Июн-19 20:39
     if (content contains "class=\"dl-unregistered") {
-      TorrentParseResult.NotAvailable("absorbed")
+      TorrentParseResult.Failure.Absorbed
     } else {
       val body = parseElement(content) \ "body"
       val downloadTable = body \\ "table" filterByClass "attach"